Adding berkshelf.config_path option
To use, add the key config.berkshelf.config_path with the value set to the path of the berkshelf config.json file to use. If this is specified, default values for client_key and node_name will be loaded from that config file. These defaults will be overriden by setting them individually.